move & remove peripheral schema classes py file into api/peripherals.py
